requirements.txt
amqp==5.0.5
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-configurations==2.2
django-cors-headers==3.7.0
django-dynamic-fixture==3.1.1
django-environ==0.4.5
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.9
phonenumbers==8.12.18
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
